pyproject.toml
[project]
name = "fastapi-starter"
version = "0.1.0"
description = "Opinionated FastAPI web service starter with Uvicorn, Typer CLI, JWT auth, tests, linting, and Docker."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Your Name", email="you@example.com"}]
dependencies = [
    "fastapi>=0.110.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "typer>=0.12.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0.12",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E","F","I","UP","B","SIM","S","N","A","C90"]
lint.ignore = ["S101"] # allow assert in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true
plugins = []
